New pileup collector using SSE. Fixed some warnings. JIRA: CXX-9613
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@79606 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/sra/readers/bam/bgzf.cpp
@@ -197,7 +197,7 @@ void CPagedFile::x_ReadPage(CPagedFilePage& page, TFilePos file_pos)
                     dst += cnt;
                 }
             }
-            double bytes = size-rem;
+            Uint8 bytes = size-rem;
             double seconds = sw.Elapsed();
             x_AddReadStatistics(bytes, seconds);
             if ( s_GetDebug() >= 5 ) {
